feat(ConvenienceController): 존재하지 않는 상품 입력 시 에러 발생시키는 기능 구현

+++ src/main/java/store/domain/MemberShipDiscountPolicy.java
@@ -3,11 +3,11 @@ package store.domain;
 public class MemberShipDiscountPolicy {
     public static final int MAXIMUM_DISCOUNT_PRICE = 8000;
 
-    public int discountAmount(int promotionTotalPrice, int promotionDiscountPrice){
-        return (int) ((promotionTotalPrice+promotionDiscountPrice) * 0.3);
+    public int discountAmount(int promotionTotalPrice, int promotionDiscountPrice) {
+        return (int) ((promotionTotalPrice + promotionDiscountPrice) * 0.3);
     }
 
-    public boolean isExceedMembershipDiscountLimit(double discountAmount){
+    public boolean isExceedMembershipDiscountLimit(int discountAmount) {
         return discountAmount > MAXIMUM_DISCOUNT_PRICE;
     }
